+++ tests/args.rs
@@ -32,7 +32,7 @@ fn negative_interval() {
     Command::cargo_bin(APP_NAME).unwrap()
         .arg("-k").arg("mock")
         .arg("-w").arg("OwmMock")
-        .arg("-l").arg("IpMock")
+        .arg("-l").arg("IpApiMock")
         .arg("-i").arg("-1000")
         .assert()
         .success();