+++ tests/common.rs
@@ -6,7 +6,7 @@ pub fn create_cmd() -> Command {
     let mut cmd = Command::cargo_bin(APP_NAME).unwrap();
     cmd.arg("-k").arg("mock")
         .arg("-w").arg("OwmMock")
-        .arg("-l").arg("IpMock")
+        .arg("-l").arg("IpApiMock")
         .arg("-i").arg("-1");
     cmd
 }