+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add &lt;temperature_feels_like&gt; tag
 - BREAKING: Replace --current-city CLI argument with &lt;city&gt; format tag
 - BREAKING: Exit after printing by specifying negative interval instead of no interval
+- BREAKING: Rename Ip to IpApi
 
 ## 1.1.1 - 2020-06-15